requirements.txt
python-telegram-bot[ext]>=20.0
google-generativeai
google-api-python-client
python-dotenv
jieba
cachetools
diskcache
Pillow
matplotlib